requirements.txt
pandas==2.2.1
pytz==2024.1
PyYAML==6.0.1
Requests==2.31.0
scikit-learn>=1.4.1
dateparser==1.2.0
genshi==0.7.7
pip~=23.2.1
configparser~=6.0.0
attrs~=23.1.0
wheel~=0.41.2
Jinja2~=3.1.2
filelock~=3.11.0
numpy~=1.24.2
setuptools~=68.1.2
parse~=1.19.1
regex~=2023.3.23
scipy~=1.11.4
threadpoolctl~=3.2.0
packaging~=23.1
torch~=2.0.0
mpmath~=1.3.0
sympy~=1.11.1
six~=1.16.0
joblib~=1.3.2
lxml~=4.9.3
MarkupSafe~=2.1.2
python-dateutil~=2.8.2
certifi~=2022.12.7
murmurhash~=1.0.9
tzlocal~=5.2
urllib3~=1.26.15
idna~=3.4